Fix: allow for unique subscription url or use rdurl origin (fixes #75)
ClusterSubscription needs just the base razeedash api url, while watch-keeper uses the razeedash api url with `/api/v2`. This change will allow the user to either specify a custom rd-api url, or will default to using the base/origin url that the watch-keeper is using. ie. the razeedash-url for watch-keeper could be `https://app.razee.io/api/v2` and the base/origin of that, for clustersubscription, would be `https://app.razee.io`.